#2845 - Improved checks for integration test

+++ src/main/scala/beam/agentsim/events/RefuelSessionEvent.scala
@@ -15,7 +15,7 @@ case class RefuelSessionEvent(
   stall: ParkingStall,
   energyInJoules: Double,
   sessionStartingFuelLevelInJoules: Double,
-  sessionDuration: Double,
+  sessionDuration: Long,
   vehId: Id[Vehicle],
   vehicleType: BeamVehicleType
 ) extends Event(tick)
